requirements.txt
streamlit
pandas
matplotlib
scikit-learn
joblib
python-dotenv
google-generativeai
numpy
